Refactor ArtistInfo to use artist image
