+++ src/components/ArtistInfo/ArtistInfo.js
@@ -3,6 +3,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { info } from '../../actions';
 import { getArtistInfo} from '../../utilities/apiCalls';
 import utils from '../../utilities/utils';
+import artistImageData from '../../utilities/artistsDataset';
 import concert from '../../assets/images/concert-stock-photo.jpeg';
 
 const ArtistInfo = ({ id }) => {
@@ -19,9 +20,22 @@ const ArtistInfo = ({ id }) => {
 
   const fetchArtistData = async (artistName) => {
     const newInfo = await getArtistInfo(artistName);
-    if (newInfo) {
-      dispatch(info(newInfo.artist));
+    const updatedInfo = findArtistImage(newInfo);
+    if (!updatedInfo.link) {
+      updatedInfo.link = concert;
     }
+    if (updatedInfo.link) {
+      dispatch(info(updatedInfo));
+    }
+  }
+
+  const findArtistImage = (newInfo) => {
+    artistImageData.forEach(artist => {
+      if (artist.name === newInfo.artist.name) {
+        newInfo.artist.link = artist.link;
+      }
+    });
+    return newInfo.artist;
   }
 
   return (
@@ -31,8 +45,10 @@ const ArtistInfo = ({ id }) => {
       </section>
       :
       <section className='artist-info fade-in' id={artistInfo.mbid}>
-        <div className='artist-img-box' style={{ backgroundImage: `url(${concert})`}}>
-          <a className='artist-name' href={artistInfo.url}>{artistInfo.name}</a>
+        <div className='artist-img-box' style={{ backgroundImage: `url(${artistInfo.link})`}}>
+          <div className='artist-image-overlay'>
+            <a className='artist-name' href={artistInfo.url}>{artistInfo.name}</a>
+          </div>
         </div>
         <div className='artist-text-box'>
           <h3 className='artist-bio-title'>Artist Biography</h3>
